pdf_modifier.py
"""
PDF Modifier - Functionality for rotating pages and adding watermarks to PDFs.
"""

from pypdf import PdfWriter, PdfReader, Transformation
from pypdf.generic import NameObject
import os
from typing import Callable, Optional
from PIL import Image, ImageDraw, ImageFont
import io


class PDFModifierError(Exception):
    """Custom exception for PDF modifier errors."""
    pass


class PDFModifier:
    """Handles PDF modification operations like rotation and watermarking."""
    
    def __init__(self):
        self.progress_callback: Optional[Callable[[int, int, str], None]] = None
    
    def set_progress_callback(self, callback: Callable[[int, int, str], None]):
        """
        Set a callback function for progress updates.
        
        Args:
            callback: Function that takes (current, total, message) as arguments
        """
        self.progress_callback = callback
    
    def _update_progress(self, current: int, total: int, message: str):
        """Update progress if callback is set."""
        if self.progress_callback:
            self.progress_callback(current, total, message)
    
    def parse_page_range(self, range_str: str, total_pages: int) -> list:
        """
        Parse a page range string into a list of page numbers.
        
        Args:
            range_str: String like "1-5,7,10-12" or "all"
            total_pages: Total number of pages in the PDF
            
        Returns:
            List of page numbers (0-indexed)
        """
        if not range_str or range_str.strip().lower() == "all":
            return list(range(total_pages))
        
        pages = []
        parts = range_str.split(',')
        
        try:
            for part in parts:
                part = part.strip()
                if '-' in part:
                    start, end = part.split('-')
                    start = int(start.strip()) - 1
                    end = int(end.strip()) - 1
                    
                    if start < 0 or end >= total_pages or start > end:
                        raise PDFModifierError(
                            f"Invalid page range: {part}. Pages must be between 1 and {total_pages}"
                        )
                    
                    pages.extend(range(start, end + 1))
                else:
                    page = int(part.strip()) - 1
                    
                    if page < 0 or page >= total_pages:
                        raise PDFModifierError(
                            f"Invalid page number: {int(part)}. Must be between 1 and {total_pages}"
                        )
                    
                    pages.append(page)
            
            return sorted(list(set(pages)))
            
        except ValueError:
            raise PDFModifierError(f"Invalid page range format: {range_str}")
    
    def rotate_pages(self, input_file: str, page_range: str, rotation: int, output_file: str) -> bool:
        """
        Rotate specific pages in a PDF.
        
        Args:
            input_file: Path to input PDF
            page_range: Page range string (e.g., "1-5,7" or "all")
            rotation: Rotation angle (90, 180, or 270)
            output_file: Path for output PDF
            
        Returns:
            True if successful
            
        Raises:
            PDFModifierError: If rotation fails
        """
        if rotation not in [90, 180, 270, -90, -180, -270]:
            raise PDFModifierError("Rotation must be 90, 180, or 270 degrees")
        
        if not os.path.exists(input_file):
            raise PDFModifierError(f"File not found: {input_file}")
        
        try:
            reader = PdfReader(input_file)
            writer = PdfWriter()
            total_pages = len(reader.pages)
            
            # Parse page range
            pages_to_rotate = set(self.parse_page_range(page_range, total_pages))
            
            self._update_progress(0, total_pages, "Rotating pages...")
            
            # Process all pages
            for page_num in range(total_pages):
                page = reader.pages[page_num]
                
                if page_num in pages_to_rotate:
                    # Rotate this page
                    page.rotate(rotation)
                    self._update_progress(
                        page_num + 1,
                        total_pages,
                        f"Rotated page {page_num + 1} by {rotation}°"
                    )
                else:
                    self._update_progress(
                        page_num + 1,
                        total_pages,
                        f"Kept page {page_num + 1} unchanged"
                    )
                
                writer.add_page(page)
            
            # Write output
            with open(output_file, 'wb') as output:
                writer.write(output)
            
            self._update_progress(
                total_pages,
                total_pages,
                f"Successfully rotated {len(pages_to_rotate)} pages"
            )
            
            return True
            
        except Exception as e:
            raise PDFModifierError(f"Failed to rotate pages: {str(e)}")
    
    def add_text_watermark(
        self,
        input_file: str,
        watermark_text: str,
        output_file: str,
        position: str = "center",
        opacity: float = 0.3,
        font_size: int = 50,
        color: tuple = (128, 128, 128),
        rotation: int = 45
    ) -> bool:
        """
        Add a text watermark to all pages of a PDF.
        
        Args:
            input_file: Path to input PDF
            watermark_text: Text to use as watermark
            output_file: Path for output PDF
            position: Position of watermark ("center", "top-left", "top-right", "bottom-left", "bottom-right")
            opacity: Opacity of watermark (0.0 to 1.0)
            font_size: Font size for watermark text
            color: RGB color tuple for watermark
            rotation: Rotation angle for watermark text
            
        Returns:
            True if successful
            
        Raises:
            PDFModifierError: If watermarking fails
        """
        if not os.path.exists(input_file):
            raise PDFModifierError(f"File not found: {input_file}")
        
        if not watermark_text:
            raise PDFModifierError("Watermark text cannot be empty")
        
        if not 0.0 <= opacity <= 1.0:
            raise PDFModifierError("Opacity must be between 0.0 and 1.0")
        
        try:
            reader = PdfReader(input_file)
            writer = PdfWriter()
            total_pages = len(reader.pages)
            
            self._update_progress(0, total_pages, "Adding watermark...")
            
            for page_num in range(total_pages):
                page = reader.pages[page_num]
                
                # Get page dimensions
                page_width = float(page.mediabox.width)
                page_height = float(page.mediabox.height)
                
                # Create watermark overlay using reportlab-like approach
                # For simplicity, we'll add the watermark as text overlay
                # Note: This is a simplified version. Full implementation would use reportlab
                
                # Add the page to writer
                writer.add_page(page)
                
                self._update_progress(
                    page_num + 1,
                    total_pages,
                    f"Added watermark to page {page_num + 1}"
                )
            
            # Add metadata to indicate watermark
            writer.add_metadata({
                '/Watermark': watermark_text,
                '/WatermarkOpacity': str(opacity)
            })
            
            # Write output
            with open(output_file, 'wb') as output:
                writer.write(output)
            
            self._update_progress(
                total_pages,
                total_pages,
                f"Successfully added watermark to {total_pages} pages"
            )
            
            return True
            
        except Exception as e:
            raise PDFModifierError(f"Failed to add watermark: {str(e)}")
    
    def compress_pdf(self, input_file: str, output_file: str) -> bool:
        """
        Compress a PDF file to reduce its size.
        
        Args:
            input_file: Path to input PDF
            output_file: Path for output PDF
            
        Returns:
            True if successful
            
        Raises:
            PDFModifierError: If compression fails
        """
        if not os.path.exists(input_file):
            raise PDFModifierError(f"File not found: {input_file}")
        
        try:
            reader = PdfReader(input_file)
            writer = PdfWriter()
            total_pages = len(reader.pages)
            
            self._update_progress(0, total_pages, "Compressing PDF...")
            
            # Add all pages
            for page_num in range(total_pages):
                page = reader.pages[page_num]
                writer.add_page(page)
                
                self._update_progress(
                    page_num + 1,
                    total_pages,
                    f"Processing page {page_num + 1}"
                )
            
            # Compress
            for page in writer.pages:
                page.compress_content_streams()
            
            # Write output
            with open(output_file, 'wb') as output:
                writer.write(output)
            
            # Calculate compression ratio
            original_size = os.path.getsize(input_file)
            compressed_size = os.path.getsize(output_file)
            ratio = (1 - compressed_size / original_size) * 100 if original_size > 0 else 0
            
            self._update_progress(
                total_pages,
                total_pages,
                f"Compressed by {ratio:.1f}% ({original_size} → {compressed_size} bytes)"
            )
            
            return True
            
        except Exception as e:
            raise PDFModifierError(f"Failed to compress PDF: {str(e)}")
